feat: add base build
