+++ .env
@@ -1,0 +1,2 @@
+HTTP_PORT=3000
+WEBSOCKET_PORT=3100
