+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "target": "ES6",
+    "baseUrl": "./",
+    "typeRoots": ["node_modules/@types"],
+    "module": "ES2020",
+    "outDir": "dist",
+    "sourceMap": false,
+    "moduleResolution": "node",
+    "noUnusedLocals": true,
+    "esModuleInterop": true
+  },
+  "include": ["src/**/*"],
+  "exclude": ["src/**/*.d.ts"]
+}
